Adding build_weapons and updating main file
build_weapons is a helper file that creates a dictionary for each class that contains the relevant information for each weapon. at >1000 lines, it was unfeasible to have it cluttering up main, and so has been moved outside for convenience. TK: i still have to verify all the numbers for the spec_A's, spec_B's, and special points are correct, but the subs, specials, ranges, weights, roles are up to date with the current weapon roster.