Added ability to update existing sheet
Added option 'C' to update an existing sheet, which has to exist in the same directory as main. Moved work to helper file as included. Assumes sheet with numbers is named "numbers", weapon names are all lowercase and in column B, and current points values are stored in column F.